+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 dev:
   - fix incorrect error when "deposit veirfy" is not given a withdrawal address
   - allow truncated mnemonics (first four characters of each word)
-  - add deneb informatoin to "block info"
+  - add deneb information to "block info"
+  - add epoch parameter to "validator yield"
 
 1.31.0:
   - initial support for deneb